pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dictionary-and-metadata-service"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "jain-kb-common",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "asyncpg>=0.29",
    "pydantic>=2.0",
    "psycopg[binary]>=3.1",
    "selectolax>=0.3",
    "PyYAML>=6.0",
    "jsonschema>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.10",
    "anyio[trio]",
]

[tool.hatch.workspace]
packages = ["packages/jain_kb_common"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
strict = true
